First commit of woodland-beacon: an implementation based on soundworks-woodland, where the ibeacon service is used for inter-device distance estimations, themselves used to estimate inter-devices IR (taps).
This first version simulates flutter effect between device pairs.
a client becomes launcher when its screen is touched, launches a
message to the server that's propagated to all clients to trigger
a play of precomputed "launcher-to-others" flutter IR (as if a
ping-pong ball was launched from launcher to every client
simultaneously, bouncing back and forth in between each individual
launcher-client pair).